(improvement)(headless)Improve dataset time query judgment (#1395)
 (improvement)(headless)Improve dataset time query judgment
The official document describes that if you do not want to add a time limit, please set the time to -1, but it has not taken effect. During LLM parsing, the parameters passed to the model still default to the date parameter. Now change to include the date parameter only when [unit!=-1 '] is checked in the dataset query settings